Super class extracted.
A super class UndoRedoAction.java, has been extracted form UndoAction.java and RedoAction.java to define common behaviour. This involves the introduction of a new method to define id and pass them to the super constructor to include in inherited behaviour.
